Make modifiers use new Fragment
