+++ Etabs_Adapter/CRUD/Create/SurfaceProperty.cs
@@ -20,21 +20,12 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using System.Collections.Generic;
-using System.Linq;
-using BH.oM.Geometry.SettingOut;
-using BH.oM.Structure.Elements;
-using BH.oM.Structure.SectionProperties;
-using BH.oM.Structure.Constraints;
+
 using BH.oM.Structure.SurfaceProperties;
-using BH.oM.Structure.Loads;
-using BH.oM.Structure.Offsets;
 using BH.Engine.Structure;
-using BH.Engine.Geometry;
-using BH.oM.Structure.MaterialFragments;
-using BH.Engine.ETABS;
-using BH.oM.Adapters.ETABS.Elements;
 using BH.oM.Adapters.ETABS;
+using BH.oM.Structure.Fragments;
+using BH.Engine.Base;
 
 #if Debug17 || Release17
 using ETABSv17;
@@ -94,11 +85,11 @@ namespace BH.Adapter.ETABS
                     retA = m_model.PropArea.SetSlab(propertyName, eSlabType.Slab, shellType, property2d.Material.DescriptionOrName(), constantThickness.Thickness);
             }
 
-
-            if (property2d.HasModifiers())
+            SurfacePropertyModifier modifier = property2d.FindFragment<SurfacePropertyModifier>();
+            if (modifier != null)
             {
-                double[] modifier = property2d.Modifiers();//(double[])property2d.CustomData["Modifiers"];
-                m_model.PropArea.SetModifiers(propertyName, ref modifier);
+                double[] modifiers = new double[] { modifier.FXX, modifier.FYY, modifier.FXY, modifier.MXX, modifier.MYY, modifier.MXY, modifier.VXZ, modifier.VYZ, modifier.Mass, modifier.Weight };
+                m_model.PropArea.SetModifiers(propertyName, ref modifiers);
             }
 
             if (retA != 0)
